Improve RadioGroup, make less buggy
Thanks to https://github.com/chenglou/react-radio-group/pull/13
